+++ src/features/chapter/utils/ChapterList.util.tsx
@@ -122,12 +122,29 @@ export const isFilterActive = (options: ChapterListFilterOptions) => {
 };
 
 export const useChapterListOptions = (manga: MangaIdInfo & GqlMetaHolder): ChapterListOptions => {
-    const { unread, downloaded, bookmarked, reverse, sortBy, showChapterNumber, excludedScanlators } =
-        useGetMangaMetadata(manga);
+    const {
+        unread,
+        downloaded,
+        bookmarked,
+        reverse,
+        sortBy,
+        showChapterNumber,
+        skipDecimalChapters,
+        excludedScanlators,
+    } = useGetMangaMetadata(manga);
 
     return useMemo(
-        () => ({ unread, downloaded, bookmarked, reverse, sortBy, showChapterNumber, excludedScanlators }),
-        [unread, downloaded, bookmarked, reverse, sortBy, showChapterNumber, excludedScanlators],
+        () => ({
+            unread,
+            downloaded,
+            bookmarked,
+            reverse,
+            sortBy,
+            showChapterNumber,
+            skipDecimalChapters,
+            excludedScanlators,
+        }),
+        [unread, downloaded, bookmarked, reverse, sortBy, showChapterNumber, skipDecimalChapters, excludedScanlators],
     );
 };