+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Added
 
 - (**Download**) Add button to retry all failed downloads
+- (**Download/Reader**) Add option to skip decimal chapters (e.g. 10.5) for download selection and the reader's next chapter navigation in case their chapter (e.g. 10) exists (global setting in the download settings, overridable per entry via its chapter list options)
 
 ### Changed